# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Vidya Socratic Tutor API
  description: |
    AI-powered Socratic tutoring API for STEM subjects and college essay coaching.
    
    Vidya uses the Socratic method to guide students toward understanding through
    carefully crafted questions — never giving direct answers or writing for the student.
    
    ## Authentication
    
    All API requests require a Bearer token (API key):
    ```
    Authorization: Bearer vk_live_xxxxxxxxxxxxxxxx
    ```
    
    API keys can be obtained by contacting the Vidya team.
    
    ## Rate Limits
    
    Rate limits are per-API-key and depend on your tier:
    - **Free**: 100 requests/minute — **hard limit** (429 when exceeded)
    - **Standard**: 500 requests/minute — **soft limit** (overage allowed and billed)
    - **Premium**: 1000 requests/minute — **soft limit** (overage allowed and billed)
    
    Standard and Premium tiers use a soft limit: requests above the plan
    limit are allowed but counted as overage and billed via Stripe metered
    billing. Free tier requests that exceed the limit receive a 429.
    
    Rate limit headers are included in every response:
    - `X-RateLimit-Limit`: Max requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Overage`: `true` when the request was allowed over the plan limit (Standard/Premium only)
    - `Retry-After`: Seconds to wait (only on 429 responses, Free tier only)
    
...
    
...
  version: 1.0.0
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /api/tutor/session/start:
    post:
      summary: Start a tutoring session
      description: |
        Creates a new tutoring session and generates the initial AI response.
        
        For essay mode (`ESSAY_WRITING`), provide `essayType`, `essayPromptText`, 
        and optionally `wordLimit` to give the AI essay-specific context.
      tags:
        - Tutoring
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartSessionRequest'
            examples:
              essay:
                summary: Start an essay coaching session
                value:
                  subject: ESSAY_WRITING
                  language: EN
                  problemText: "I need help with my Common App personal statement about overcoming a challenge."
                  essayType: personal_statement
                  essayPromptText: "Some students have a background, identity, interest, or talent that is so meaningful they believe their application would be incomplete without it."
                  wordLimit: 650
              stem:
                summary: Start a STEM tutoring session
                value:
                  subject: PHYSICS
                  language: EN
                  problemText: "A ball is thrown vertically upward with a velocity of 20 m/s. Find the maximum height reached."
      responses:
        '200':
          description: Session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartSessionResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/tutor/message:
    post:
      summary: Send a message in a session
      description: |
        Sends a student message and receives an AI Socratic response.
        
        For essay mode, the student should paste their essay draft as the message.
        The AI will analyze the draft and respond with guiding questions.
      tags:
        - Tutoring
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
            examples:
              essay_draft:
                summary: Student submits essay draft
                value:
                  sessionId: "clxyz123abc"
                  message: "Here's my draft: Growing up in a small town, I learned that challenges are not obstacles but stepping stones..."
              stem_attempt:
                summary: Student attempts STEM problem
                value:
                  sessionId: "clxyz123abc"
                  message: "I think I need to use the equation v² = u² - 2gh, so h = u²/2g = 400/20 = 20m"
      responses:
        '200':
          description: AI response generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendMessageResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Session not found
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/tutor/session/{sessionId}:
    get:
      summary: Get session details
      description: Returns the session metadata and full message history.
      tags:
        - Tutoring
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetSessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Session not found
  /api/tutor/session/{sessionId}/end:
    post:
      summary: End a tutoring session
      description: |
        Marks the session as completed, generates a structured coaching report
        using AI, populates learning analytics (conceptsUsed, masteryGain, resolved),
        and returns the full report.
        
        If the session has already been ended, returns the existing report.
        The report generation uses Haiku (cheap, ~$0.001 per call).
      tags:
        - Tutoring
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session ended with report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndSessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Session not found
  /api/tutor/session/{sessionId}/summary:
    get:
      summary: Get session report/summary
      description: |
        Returns the session coaching report generated when the session was ended.
        Returns 404 if the session hasn't been ended yet (no report available).
      tags:
        - Tutoring
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session report
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  session:
                    type: object
                    properties:
                      id:
                        type: string
                      subject:
                        $ref: '#/components/schemas/Subject'
                      status:
                        $ref: '#/components/schemas/SessionStatus'
                      resolved:
                        type: boolean
                      masteryGain:
                        type: number
                        nullable: true
                      conceptsUsed:
                        type: array
                        items:
                          type: string
                      startedAt:
                        type: string
                        format: date-time
                      endedAt:
                        type: string
                        format: date-time
                        nullable: true
                  report:
                    $ref: '#/components/schemas/SessionReport'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Session not found or report not yet generated
  /api/tutor/session/{sessionId}/quiz:
    post:
      summary: Generate a micro-quiz from a session
      description: |
        Returns a short quiz (default 3 questions) based on the session report.
      tags:
        - Tutoring
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizRequest'
      responses:
        '200':
          description: Session quiz
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Session not found
  /api/tutor/sessions:
    get:
      summary: List sessions
      description: |
        Returns a list of sessions, optionally filtered by user, subject, or status.
        Each session includes learning analytics (conceptsUsed, masteryGain, resolved)
        and the session report if the session has been ended.
        Sessions are scoped to the authenticated API key.
      tags:
        - Tutoring
      parameters:
        - name: userId
          in: query
          schema:
            type: string
        - name: subject
          in: query
          schema:
            $ref: '#/components/schemas/Subject'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/SessionStatus'
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 100
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListSessionsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================
  # DEVELOPER PLATFORM
  # ============================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key obtained from the Vidya admin panel
  schemas:
...
    StartSessionRequest:
      type: object
      required:
        - subject
        - problemText
      properties:
        userId:
          type: string
          description: External user ID (your system's user identifier)
        subject:
          $ref: '#/components/schemas/Subject'
        language:
          $ref: '#/components/schemas/Language'
          default: EN
        problemText:
          type: string
          minLength: 1
          maxLength: 5000
          description: The problem or essay prompt to work on
        problemImage:
          type: string
          description: Image reference (http/https URL or data:image/* base64 string)
        noFinalAnswer:
          type: boolean
          description: When true, the tutor will avoid final answers and stay Socratic
        essayType:
          type: string
          description: Type of essay (e.g., personal_statement, why_us, activity)
        essayPromptText:
          type: string
          maxLength: 10000
          description: The full essay prompt text from the application
        wordLimit:
          type: integer
          minimum: 1
          description: Word limit for the essay
...
    StartSessionResponse:
      type: object
      properties:
        success:
          type: boolean
        session:
          type: object
          properties:
            id:
              type: string
            subject:
              $ref: '#/components/schemas/Subject'
            language:
              $ref: '#/components/schemas/Language'
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
    SendMessageResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          $ref: '#/components/schemas/Message'
...
  responses:
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: Validation error
            details:
              - path: ["subject"]
                message: "Required"
...